requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.34.0
openai==1.3.7
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
PyMuPDF==1.23.8
Pillow==10.1.0
